# meal-planner-system-frontend/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getAll_1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mealType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BREAKFAST",
                                "LUNCH",
                                "DINNER",
                                "DESSERT",
                                "DRINK",
                                "SNACK",
                                "SAUCE_OR_CONDIMENT",
                                "UNCLASSIFIED"
                            ]
                        }
                    },
                    {
                        "name": "cookTime",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "MIN_15",
                                "MIN_30",
                                "MIN_60",
                                "HOURS_4",
                                "DAYS_1_PLUS"
                            ]
                        }
                    },
                    {
                        "name": "cookComplexity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "EASY",
                                "MEDIUM",
                                "HARD"
                            ]
                        }
                    },
                    {
                        "name": "cookBudget",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LOW",
                                "MEDIUM",
                                "HIGH"
                            ]
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "default": [
                                "name,ASC"
                            ],
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedModelRecipeSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getById_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes/slug/{slug}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes/by-ingredient/{ingredientId}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getByIngredient",
                "parameters": [
                    {
                        "name": "ingredientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedModelRecipeSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ingredients/slug/{slug}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getBySlug_1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/categories/slug/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "getBySlug_2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IngredientResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "normalizedName": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "CategoryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "parentName": {
                        "type": "string"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResponse"
                        }
                    }
                }
            },
...
            "PageMetadata": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "number": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "PagedModelRecipeSummaryResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeSummaryResponse"
                        }
                    },
                    "page": {
                        "$ref": "#/components/schemas/PageMetadata"
                    }
                }
            },
            "RecipeSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "mealType": {
                        "type": "string"
                    },
                    "cookTime": {
                        "type": "string"
                    },
                    "cookComplexity": {
                        "type": "string"
                    },
                    "cookBudget": {
                        "type": "string"
                    },
                    "servings": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "RecipeIngredientResponse": {
                "type": "object",
                "properties": {
                    "ingredientId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rawName": {
                        "type": "string"
                    },
                    "rawAmount": {
                        "type": "string"
                    },
                    "availableInShop": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "imageUrl": {
                        "type": "string"
                    }
                }
            },
            "RecipeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "mealType": {
                        "type": "string"
                    },
                    "mealTypeDetailed": {
                        "type": "string"
                    },
                    "cookTime": {
                        "type": "string"
                    },
                    "cookComplexity": {
                        "type": "string"
                    },
                    "cookBudget": {
                        "type": "string"
                    },
                    "servings": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "servingSize": {
                        "type": "string"
                    },
                    "steps": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeIngredientResponse"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}